E2020 errors

+++ Project/DuckEngine/Engine/src/API_Test.cpp
@@ -89,7 +89,7 @@ void Application::startScriptEngine()
     (
         runtimePath.c_str(),     // AppDomain base path
         "SampleHost",            // AppDomain friendly name, this can be anything you want really
-        propertyKeys.size(),     // Property count
+        static_cast<int>(propertyKeys.size()),     // Property count
         propertyKeys.data(),     // Property names
         propertyValues.data(),   // Property values
         &hostHandle,             // Host handle
